requirements.txt
flask
flask_cors
tqdm
python-dotenv
flask-sqlalchemy
pytest
requests
flask_swagger_ui